feat(mpsc): stick errors in their own module
